requirements.txt
fastapi
uvicorn
transformers
chromadb
torch
pillow
tqdm
